Strip back unused code
Google and GitHub OAuthentication are being removed because without the
add_user hooks they are exactly they same as the jupyter/oauthenticator
versions (which we need as a dependency anyway). pam.py is being removed
for the same reason. I've made a guess at the modifications to
swiftspawner but I don't know enough about it to evaluate it yet.

Signed-off-by: Callysto Sysadmin <sysadmin@callysto.ca>